requirements.txt
Django==1.11
djangorestframework
